[fix] move setFCM() function to mainActivity from signActivity
