+++ app/src/main/java/org/gdsc/donut/ui/GiverMainActivity.kt
@@ -1,9 +1,21 @@
 package org.gdsc.donut.ui
 
+import android.Manifest
+import android.content.ContentValues
+import android.content.pm.PackageManager
+import android.os.Build
 import android.os.Bundle
+import android.util.Log
 import android.view.View
+import android.widget.Toast
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.content.ContextCompat
+import com.google.android.gms.tasks.OnCompleteListener
+import com.google.firebase.messaging.FirebaseMessaging
 import org.gdsc.donut.R
+import org.gdsc.donut.data.DonutSharedPreferences
 import org.gdsc.donut.databinding.ActivityGiverMainBinding
 import org.gdsc.donut.ui.donation.DonationCheckFragment
 import org.gdsc.donut.ui.donation.DonationFragment
@@ -14,9 +26,15 @@ import org.gdsc.donut.ui.home.GiverWalletFragment
 import org.gdsc.donut.ui.home.WalletDetailFragment
 import org.gdsc.donut.ui.mypage.GiverMyPageFragment
 import org.gdsc.donut.ui.ranking.RankingFragment
+import org.gdsc.donut.ui.viewModel.SignViewModel
 
 class GiverMainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityGiverMainBinding
+    private val viewModel: SignViewModel by viewModels()
+    private val requestPermissionLauncher = registerForActivityResult(ActivityResultContracts.RequestPermission(),) { isGranted: Boolean ->
+        if (isGranted) setFCM()
+        else Toast.makeText(baseContext, "Turn off push notification", Toast.LENGTH_SHORT).show()
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -25,6 +43,7 @@ class GiverMainActivity : AppCompatActivity() {
 
         setBottomNavigation()
         setFloatingButton()
+        askNotificationPermission()
     }
 
     override fun onBackPressed() {
@@ -90,5 +109,26 @@ class GiverMainActivity : AppCompatActivity() {
 
     fun enableFloatingButton(){
         binding.fabDonationBtn.visibility = View.VISIBLE
+    }
+
+    private fun askNotificationPermission() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            if (ContextCompat.checkSelfPermission(this, Manifest.permission.POST_NOTIFICATIONS) == PackageManager.PERMISSION_GRANTED) {
+                setFCM()
+            } else {
+                requestPermissionLauncher.launch(Manifest.permission.POST_NOTIFICATIONS)
+            }
+        }
+    }
+
+    private fun setFCM(){
+        FirebaseMessaging.getInstance().token.addOnCompleteListener(OnCompleteListener { task ->
+            if (!task.isSuccessful) {
+                Log.w(ContentValues.TAG, "Fetching FCM registration token failed", task.exception)
+                return@OnCompleteListener
+            }
+
+            DonutSharedPreferences.getAccessToken()?.let { viewModel.sendFCMToken(it, task.result)}
+        })
     }
 }
